+++ src/bmlite/SPM/_simulation.py
@@ -17,8 +17,10 @@ if TYPE_CHECKING:  # pragma: no cover
 
 class Simulation(object):
 
-    __slots__ = ['_yamlfile', '_yamlpath', '_flags', '_t0', '_sv0', '_svdot0',
-                 '_lband', '_uband', '_algidx', 'bat', 'el', 'an', 'ca']
+    __slots__ = [
+        '_yamlfile', '_yamlpath', '_flags', '_t0', '_sv0', '_svdot0', '_lband',
+        '_uband', '_algidx', 'c', 'bat', 'el', 'an', 'ca',
+    ]
 
     def __init__(self, yamlfile: str = 'graphite_nmc532') -> None:
         """
@@ -56,6 +58,7 @@ class Simulation(object):
 
         """
 
+        from .. import Constants
         from .domains import Battery, Electrolyte, Electrode
 
         if '.yaml' not in yamlfile:
@@ -80,6 +83,7 @@ class Simulation(object):
         yaml = YAML(typ='safe')
         yamldict = yaml.load(yamlpath)
 
+        self.c = Constants()
         self.bat = Battery(**yamldict['battery'])
         self.el = Electrolyte(**yamldict['electrolyte'])
         self.an = Electrode('anode', **yamldict['anode'])
@@ -149,42 +153,60 @@ class Simulation(object):
         self._lband = int(self.el.ptr['phi_el'] - self.an.r_ptr['Li_ed'][-1])
         self._uband = int(self.el.ptr['phi_el'] - self.an.r_ptr['Li_ed'][-1])
 
-    def j_pattern(self) -> None:
+    def j_pattern(self, plot: bool = True,
+                  return_bands: bool = False) -> tuple[int] | None:
         """
-        Plot the Jacobian pattern.
+        Determine the Jacobian pattern.
 
-        Runs the ``bmlite.SPM.dae.bandwidth`` function to determine and plot
-        the Jacobian pattern.
+        Parameters
+        ----------
+        plot : bool, optional
+            Whether or not to plot the Jacobian pattern. The default is True.
+        return_barnds : bool, optional
+            Whether or not to return the half bandwidths (lower, upper). The
+            default is False.
 
         Returns
         -------
         lband : int
-            Lower bandwidth from the residual function's Jacobian pattern.
+            The lower half bandwidth. Only returned if `return_bands=True`.
         uband : int
-            Upper bandwidth from the residual function's Jacobian pattern.
-
-        See also
-        --------
-        bmlite.SPM.dae.bandwidth
+            The upper half bandwidth. Only returned if `return_bands=True`.
 
         """
 
-        from .dae import bandwidth
+        from .dae import residuals
+        from .._core._idasolver import bandwidth
         from bmlite.plotutils import format_ticks, show
 
-        lband, uband, j_pat = bandwidth(self)
+        t0 = 0.
+        y0 = np.hstack([self.an.sv0(), self.el.sv0(), self.ca.sv0()])
+        yp0 = np.zeros_like(y0)
 
-        fig, ax = plt.subplots(nrows=1, ncols=1, figsize=[4, 4],
-                               layout='constrained')
+        step = {
+            'mode': 'current',
+            'units': 'C',
+            'value': lambda t: 0.,
+        }
 
-        ax.spy(j_pat)
-        ax.text(0.1, 0.2, 'lband: ' + str(lband), transform=ax.transAxes)
-        ax.text(0.1, 0.1, 'uband: ' + str(uband), transform=ax.transAxes)
+        userdata = (self, step)
 
-        format_ticks(ax)
-        show(fig)
+        lband, uband, j_pat = bandwidth(residuals, t0, y0, yp0, userdata,
+                                        return_pattern=True)
 
-        return lband, uband
+        if plot:
+            fig, ax = plt.subplots(nrows=1, ncols=1, figsize=[4, 4],
+                                layout='constrained')
+
+            ax.spy(j_pat)
+            ax.text(0.1, 0.2, 'lband: ' + str(lband), transform=ax.transAxes)
+            ax.text(0.1, 0.1, 'uband: ' + str(uband), transform=ax.transAxes)
+
+            format_ticks(ax)
+            show(fig)
+
+        if return_bands:
+            return lband, uband
 
     def run_step(self, expr: Experiment, stepidx: int) -> StepSolution:
         """